laconia#48 - serverless-offline raises warning due to callback usage; commit: testing laconia without callback
